Large restructure of differential phase contrast class structure
- All the DPCSignal functionality has been moved to the BeamShift class
- BeamShift.plot() default plotting has been changed
- Some unit tests with random numbers which were prone to failing
randomly has been removed
- Some rarely used functions has been removed
